Save High Reach content identity and storage codec work

+++ src/systems/world/geometry.ts
@@ -11,10 +11,10 @@ import { isShaped, getShapeBoxes } from './blockShapes';
 import { getAtlasDimensions, ATLAS_RAW_TILE_SIZE, ATLAS_PADDING, ATLAS_STRIDE } from '../../utils/textures';
 
 export interface NeighborData {
-    left?: Uint8Array;
-    right?: Uint8Array;
-    front?: Uint8Array;
-    back?: Uint8Array;
+    left?: Uint16Array;
+    right?: Uint16Array;
+    front?: Uint16Array;
+    back?: Uint16Array;
 }
 
 export interface NeighborLight {
@@ -221,7 +221,7 @@ const SHAPED_FACES: { name: 'right' | 'left' | 'top' | 'bottom' | 'front' | 'bac
 export function generateGeometryData(
     _cx: number,
     _cz: number,
-    chunk: Uint8Array,
+    chunk: Uint16Array,
     metaData: Uint8Array | undefined,
     neighbors: NeighborData,
     lights: NeighborLight,
